restructure sponsors
