+++ src/pages/_app.tsx
@@ -1,5 +1,5 @@
 import "@/styles/globals.css";
-import "./landingpage/styles/Header.css"
+import "./landingpage/styles/Header.css";
 import type { AppProps } from "next/app";
 export default function App({ Component, pageProps }: AppProps) {
   return (
@@ -9,7 +9,7 @@ export default function App({ Component, pageProps }: AppProps) {
         backgroundSize: "cover",
         backgroundRepeat: "no-repeat",
         // backgroundPosition: "center center",
-        minHeight: "100vh",
+        minHeight: "100vh",        
       }}
     >
       <Component {...pageProps} />
