+++ src/styles/Sponsor.module.css
@@ -7,9 +7,6 @@
 }
 .sponsor_bx
 {
-    background-image:url('../static/images/sponsorbg.png');
-    background-size: cover;
-    background-repeat: no-repeat;
     text-align: center;
     text-align: center;
      padding: 60px 50px;
